+++ tests/phpunit/Unit/View/SelectTest.php
@@ -53,8 +53,13 @@ class SelectTest extends AbstractViewTestCase {
 	 */
 	public function test_render__no_choices() {
 
-		$element = $this->get_element( 'element', new ArrayChoiceList( [] ) );
-		static::assertSame( '<select name="element"></select>', ( new Select() )->render( $element ) );
+		$element  = $this->get_element( 'element', new ArrayChoiceList( [] ) );
+		$rendered = ( new Select() )->render( $element );
+
+		static::assertContains( '<select', $rendered );
+		static::assertContains( 'name="element"', $rendered );
+		static::assertContains( 'id="element"', $rendered );
+		static::assertContains( '</select>', $rendered );
 	}
 
 	/**